refact : 상세 정보 / 댓글 모델 분리
- detail.model.ts에서 댓글 모델 제거
- comment.model.ts 생성
