+++ src/models/comment.model.ts
@@ -1,0 +1,8 @@
+export interface ICommentItem {
+  id: number;
+  user_email: string;
+  nickname: string;
+  comment: string;
+  updated_at: Date;
+  isMine: boolean;
+}
